python-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
azure-ai-projects==1.0.0
azure-ai-agents==1.2.0b2
azure-identity==1.15.0
azure-ai-textanalytics==5.3.0
requests==2.31.0
pydantic==2.5.0
python-multipart==0.0.6
httptools
aiohttp==3.9.1
